Added dependencies, created class shield and modified class Item,Player,Fight

+++ src/main/java/com/extraction/items/Medikit.java
@@ -3,7 +3,7 @@ package com.extraction.items;
 public class Medikit extends Item {
     private int healingAmount_;
 
-    public Medikit(ItemName name, int weight, int healingAmount) {
+    public Medikit(ItemName name, double weight, int healingAmount) {
         super(name, weight);
         this.healingAmount_ = healingAmount;
     }
